requirements.txt
aioboto3
pyyaml
requests
